refactor streaming index logic into hook
